requirements.txt
requests
eth-brownie
click
